Refactor _fix_notification_init function
