+++ test/integration/ggrc/notifications/__init__.py
@@ -1,2 +1,61 @@
 # Copyright (C) 2019 Google Inc.
 # Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
+"""Fixes Notification object init function"""
+
+import functools
+from datetime import datetime
+from sqlalchemy import and_
+
+from ggrc import db
+from ggrc.models import Notification
+from ggrc.models import NotificationType
+from integration.ggrc import TestCase
+
+
+class TestNotifications(TestCase):
+  """Base class for fixing Notification object init function"""
+
+  def _fix_notification_init(self):
+    """Fix Notification object init function.
+
+    This is a fix needed for correct created_at field when using freezgun. By
+    default the created_at field is left empty and filed by database, which
+    uses system time and not the fake date set by freezugun plugin. This fix
+    makes sure that object created in freeze_time block has all dates set with
+    the correct date and time.
+    """
+    # pylint: disable=missing-docstring
+    def init_decorator(init):
+      """"Adjust the value of the object's created_at attribute to now."""
+      @functools.wraps(init)
+      def new_init(self, *args, **kwargs):
+        init(self, *args, **kwargs)
+        if hasattr(self, "created_at"):
+          self.created_at = datetime.now()
+      return new_init
+
+    Notification.__init__ = init_decorator(Notification.__init__)
+
+  @staticmethod
+  def _get_notifications(sent=False, notif_type=None):
+    """Get a notification query.
+
+    Args:
+      sent (boolean): flag to filter out only notifications that have been
+        sent.
+      notif_type (string): name of the notification type.
+
+    Returns:
+      sqlalchemy query for selected notifications.
+    """
+    if sent:
+      notif_filter = Notification.sent_at.isnot(None)
+    else:
+      notif_filter = Notification.sent_at.is_(None)
+
+    if notif_type:
+      notif_filter = and_(notif_filter, NotificationType.name == notif_type)
+
+    return db.session.query(Notification).join(NotificationType).filter(
+        notif_filter
+    )
