+++ test/integration/ggrc/notifications/test_reminderable.py
@@ -3,20 +3,17 @@
 
 """Tests for reminders for models with reminderable mixin."""
 
-from datetime import datetime
 from freezegun import freeze_time
-from sqlalchemy import and_
 
-from ggrc import db
 from ggrc import models
 from ggrc.notifications import common
 
-from integration.ggrc import TestCase
 from integration.ggrc import api_helper
 from integration.ggrc.models import factories
+from integration.ggrc.notifications import TestNotifications
 
 
-class TestReminderable(TestCase):
+class TestReminderable(TestNotifications):
 
   """Test sending reminder."""
 
@@ -25,51 +22,6 @@ class TestReminderable(TestCase):
     self.client.get("/login")
     self._fix_notification_init()
     self.api_helper = api_helper.Api()
-
-  def _fix_notification_init(self):
-    """Fix Notification object init function.
-
-    This is a fix needed for correct created_at field when using freezgun. By
-    default the created_at field is left empty and filed by database, which
-    uses system time and not the fake date set by freezugun plugin. This fix
-    makes sure that object created in freeze_time block has all dates set with
-    the correct date and time.
-    """
-
-    def init_decorator(init):
-      """Wrapper for Notification init function."""
-
-      def new_init(self, *args, **kwargs):
-        init(self, *args, **kwargs)
-        if hasattr(self, "created_at"):
-          self.created_at = datetime.now()
-      return new_init
-
-    models.Notification.__init__ = init_decorator(models.Notification.__init__)
-
-  @classmethod
-  def _get_notifications(cls, sent=False, notif_type=None):
-    """Get a notification query.
-
-    Args:
-      sent (boolean): flag to filter out only notifications that have been
-        sent.
-      notif_type (string): name of the notification type.
-
-    Returns:
-      sqlalchemy query for selected notifications.
-    """
-    if sent:
-      notif_filter = models.Notification.sent_at.isnot(None)
-    else:
-      notif_filter = models.Notification.sent_at.is_(None)
-
-    if notif_type:
-      notif_filter = and_(notif_filter,
-                          models.NotificationType.name == notif_type)
-
-    return db.session.query(models.Notification).join(
-        models.NotificationType).filter(notif_filter)
 
   def create_assessment(self, people=None):
     """Create default assessment with some default assignees in all roles.
